test(api): migrate health checks to api-first bdd
Round 27 of the API-first BDD migration (#16967), HEALTH.

- Extend createBddApi with the health and healthAuth clients.
- Add health.bdd.test.ts: the public health check returns {status:"ok"} and the
  authenticated health check is 401 without credentials. The authenticated
  success paths remain in the auth-probe suite.
- Delete health.test.ts whole.

Coverage verified (source-only): health.ts (3/0) and health-auth-probe.ts
(18/10) unchanged vs main. No regressions.
